Fix oneline commands being mistreated.
Command sequences like

    =item [Oops]
    Problematic item here

were incorrectly concatenated into:

    =item [Oops]Problematic item here

That is, the newline was not honoured as whitespace. As a result, the
generated HTML looked like this:

    <dt>Oops]Problemati</dt><dd>item here</dd>

(the "c" went missing because the dt formatter erases the last
character, as it expects that one to be the closing "]")

This commit fixes that. Markup like the above now generates:

    <dt>Opps</dt><dd>Problematic item here</dd>

Which is what it should do.